Add #delete and #delete_node
Added a method #delete to check for a value in the BST and delete this from the BST. This is now only working for leaves and nodes with only a right or left child. #delete_node is a private method that handles the specifics for checking node.left and node.right and passing on the right return value in place of the node to be deleted.