Add #level_order, recursive and iterative version
This was a very interesting excercise. Turns out the iterative version was easier to create, though the recursive one seemed quite intuitive once the iterative one was made.

With the recursive method, I ran into the interesting problem of how to work with the block. Obviously this has to be passed on every recursion. I'm glad I tried to make the recursive method as well, as this forced me to run into the use of & to create a Proc variable.