Add #inorder that can take a block, returns array
Currently every node.left and node.right are called recursively even when nil. the recursive call simply returns if a node.data equals nil. This could be taken care of within the recursive calls themselves with an if statement, but for now this seems to be more clear and consistent behavior.