Add #rebalance and #pretty_print
Implementing #rebalance the way I have done now breaks the normal behavior of creating/building the tree. When executing tree = Tree.new(array), instead of returning the root node, there is a warning saying "Object doesn't support #inspect". For now I would say it would be preferable to be have the script return the root node upon creating the tree, but it's not breaking anything as far as it seems.

I think this requires some deeper understanding to fix. Most Google results led to issues with Ruby on Rails. So, for now I'll leave it as is.

The #pretty_print was gratefully taken from the Odin website.